Refactor duke packages
