+++ src/main/java/duke/Duke.java
@@ -6,6 +6,11 @@ import java.util.Scanner;
 
 import duke.exception.DukeFileSystemException;
 import duke.exception.InvalidDukeCommandException;
+import duke.io.DukeParser;
+import duke.io.Storage;
+import duke.io.Ui;
+import duke.task.*;
+import duke.type.DukeCommand;
 
 
 /**
