+++ src/test/java/duke/DukeParserTest.java
@@ -2,6 +2,8 @@ package duke;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import duke.io.DukeParser;
+import duke.type.DukeCommand;
 import org.junit.jupiter.api.Test;
 
 public class DukeParserTest {
