+++ src/test/java/duke/TodoTest.java
@@ -2,6 +2,7 @@ package duke;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import duke.task.Todo;
 import org.junit.jupiter.api.Test;
 
 public class TodoTest {
